+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 # Upcoming
 
 ### ✅ Added
+- Parameters to allow custom mention parsing logic. If set to `false`, `Message.mentionedUsers` is not overridden on send. [#338](https://github.com/GetStream/stream-chat-swift/issues/338)
+  - `parseMentionedUsers: Bool = true` parameter in `Client.send(message: ...)`.
+  - `parseMentionedUsers: Bool = true` parameter in `ChannelPresenter.send(text: ...)`.
+  - `parseMentionedUsersOnSend: Bool = true` property in `ChatViewController`. 
 - Message search with filter for messages [#348](https://github.com/GetStream/stream-chat-swift/pull/348)
   - `Client.search(filter: Filter, messageFilter: Filter, ...)` [#348](https://github.com/GetStream/stream-chat-swift/pull/348)
   - `SearchQuery.init(filter: Filter, messageFilter: Filter, ...)` [#348](https://github.com/GetStream/stream-chat-swift/pull/348)